update fix loi thong ke theo ngay, loi crash app khi chi nhap so tien qua nhieu

+++ app/src/main/java/com/example/doantotnghiep/adapter/AdapterKeHoachTietKiem.java
@@ -55,7 +55,7 @@ public class AdapterKeHoachTietKiem extends ArrayAdapter<ArrayKeHoachTietKiem> {
         return view;
     }
 
-    public static String DoiSoSangTien(int so) {
+    public static String DoiSoSangTien(double so) {
         DecimalFormat decimalFormat = (DecimalFormat) NumberFormat.getInstance(Locale.US);
         decimalFormat.applyPattern("#,###,###,###");
         return decimalFormat.format((so)) + " đ";
