+++ app/src/main/java/com/example/doantotnghiep/adapter/AdapterLichSuChuyenTien.java
@@ -47,7 +47,7 @@ public class AdapterLichSuChuyenTien extends ArrayAdapter<ArrayLichSuChuyenTien>
         return view;
     }
 
-    public static String DoiSoSangTien(int so) {
+    public static String DoiSoSangTien(double so) {
         DecimalFormat decimalFormat = (DecimalFormat) NumberFormat.getInstance(Locale.US);
         decimalFormat.applyPattern("#,###,###,###");
         return decimalFormat.format((so)) + " đ";
