+++ app/src/main/java/com/example/doantotnghiep/adapter/AdapterThuChiChoKeHoachTietKiem.java
@@ -47,7 +47,7 @@ public class AdapterThuChiChoKeHoachTietKiem extends ArrayAdapter<ArrayThuChiCho
         return view;
     }
 
-    public static String DoiSoSangTien(int so) {
+    public static String DoiSoSangTien(double so) {
         DecimalFormat decimalFormat = (DecimalFormat) NumberFormat.getInstance(Locale.US);
         decimalFormat.applyPattern("#,###,###,###");
         return decimalFormat.format((so)) + " đ";
